carve out MultipleClassLoaderTest as its own class
